Add Blob File Checker
In this update of main.java, the file will check to see if you have a 'Blobs' folder and create one if not in the same path as the main.java file. This makes the portability a little better. Next task is to download the latest version of TSSChecker automatically to the same location as main.java so that you can run main.java from anywhere without worrying about dependencies.