Fix Small Bug
See last update description for more information. This update fixes the printing of "done" when main.java creates a new "Blobs" folder. I implemented this to make sure that it was successfully creating the blobs folder when there wasn't one and forgot to take it out before submitting the changes.