Add Back Button and Auto Get TSSChecker
In this most recent update, I added a back button to the saved device page and to the new device page. This provides a lot more functionality and ease of usage for people.
I also added the feature to automatically get tsschecker from Tihmstar's github page.
Small bug fixes as well.